requirements.txt
aiofiles==25.1.0
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.6.3
breadability==0.1.20
cachetools==4.2.2
certifi==2025.11.12
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
distro==1.9.0
docopt==0.6.2
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.11
jiter==0.12.0
joblib==1.5.2
lxml==6.0.2
nltk==3.9.2
numpy==2.3.4
openai==2.7.2
pillow==12.0.0
PyArabic==0.6.15
pycountry==24.6.1
pydantic==2.12.4
pydantic_core==2.41.5
PyMuPDF==1.26.6
python-pptx==1.0.2
python-telegram-bot==20.6
pytz==2025.2
regex==2025.11.3
requests==2.32.5
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sumy==0.11.0
tornado==6.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
xlsxwriter==3.2.9
Flask==2.3.3